project-forge/templates/googel cloud/requirements.txt
google-cloud-bigquery==3.35.0
google-cloud-storage==3.1.1
google-cloud-pubsub==2.31.1
google-cloud-secret-manager==2.24.0
google-cloud-firestore==2.21.0
google-cloud-tasks==2.19.3
google-cloud-logging==3.12.1
google-cloud-sql-connector[pymysql]==1.18.2
sqlalchemy==2.0.40
python-dotenv==1.0.1
